Switch to a vchan-like Endpoint.Make interface
Most of the code between netfront and netback will be shared so it's
convenient to have them in the same module.

Note the API has changed: the DEVICE.id we use is now a polymorphic
variant which is either `Client or `Server. Once connected, we can
use write{,v} and listen as normal.

Note that, unlike VCHAN which is a FLOW, this is a NETWORK and a NETWORK
is not a FLOW because it has 'listen' with a callback.

Signed-off-by: David Scott <dave.scott@citrix.com>